requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==5.0.4
django-debug-toolbar==4.3.0
django-grappelli==3.0.8
django-jinja==2.11.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-tinymce==4.0.0
gunicorn==21.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.0
pillow==10.3.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
webencodings==0.5.1
